Adding a note about user Garnaat
A few of the scripts that Garnaat wrote are in this directory.  I've had
some modifications to my own (and his) scripts for years but never
pushed them.  I'm too busy right now to figure out which is which so
playing it safe, attributing, and will come back with more time to
document better.